Ticket/fixes (#103)
* 21

* 22

* 23

* 24

* 25

* 26

* 27

* 28

* 29

* 11

* 12

* 15

* 20

* 31

* 10

* 18-19

* 7, 16

* 17

* 13

* 14

* 1

* 2

* 3

* 4

* 5

* 8

* 9

* 30

* 30

* 34

* Cancel and pay buttons added for processing orders

* Email sending fix

* Tests fixes

* PHPStan implemented

* Composer fix

+++ src/database/migrations/2026_05_03_203839_create_cart_items_table.php
@@ -6,7 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -17,7 +18,7 @@ return new class extends Migration {
             $table->foreignId('cart_id')->constrained('carts')->cascadeOnDelete();
             $table->foreignId('product_id')->constrained('products')->cascadeOnDelete();
             $table->integer('quantity');
-            $table->integer('price');
+            $table->unsignedBigInteger('price');
             $table->timestamps();
         });
     }
