+++ src/database/migrations/2026_05_05_094928_create_orders_table.php
@@ -6,7 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -20,7 +21,7 @@ return new class extends Migration {
             $table->string('customer_phone')->nullable();
             $table->text('shipping_address');
             $table->string('status')->default('pending');
-            $table->integer('total_amount_cents');
+            $table->unsignedBigInteger('total_amount_cents');
             $table->timestamps();
         });
     }
