+++ src/database/migrations/2026_05_05_094938_create_order_items_table.php
@@ -6,7 +6,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -18,7 +19,7 @@ return new class extends Migration {
             $table->foreignId('product_id')->nullable()->constrained()->nullOnDelete();
             $table->string('product_name');
             $table->integer('quantity');
-            $table->integer('price_cents');
+            $table->unsignedBigInteger('price_cents');
             $table->timestamps();
         });
     }
