+++ src/database/migrations/2026_05_05_094948_create_payments_table.php
@@ -7,7 +7,8 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration {
+return new class extends Migration
+{
     /**
      * Run the migrations.
      */
@@ -19,7 +20,7 @@ return new class extends Migration {
             $table->string('provider');
             $table->string('transaction_id')->nullable()->unique();
             $table->string('status')->default(PaymentStatus::Pending);
-            $table->integer('amount_cents');
+            $table->unsignedBigInteger('amount_cents');
             $table->timestamps();
         });
     }
